AskdjangoBasic/askcompany/requirements.txt
django == 2.1.1
django-extensions==2.1.2
django-debug-toolbar==3.2.1
#jinja2
#django-jinja
pillow